Chakra UI (#93)
* chakra ui - installed

* LoadMoreBtn

* custom theme - removed

* footer

* about

* page urls

Temp solution before adding `routes.ts`

* spinner

* i18n - source file names

* logo - name

* app - layout

* helmet - page titles

* fix - removed helmet

Not a solo page. Renders in HOME.

* page - 404

* renamed - page 404

* page - noMovieFound + rows

* search bar

* LoadMoreBtn + refactoring

* header

* movieThumb + isClickable

* header - text + responsive

* SubNavigation

* MovieInfo

* Actor

* layout - todo

* Grid

* codacy - search - fix 'was used before it was defined'

Disallow Early Use (no-use-before-define) - In JavaScript, prior to ES6, variable and function declarations are hoisted to the top of a scope, so it's possible to use identifiers before their formal declarations in code. This can be confusing and some believe it is best to always declare variables and functions before using them.

* codacy - movieInfo - fix 'was used before it was defined'

Moving CSS consts before component.